refactor: refactor module dependencies
