+++ jakta-distributed/jakta-distributed-client/build.gradle.kts
@@ -25,5 +25,4 @@ dependencies {
     implementation(libs.ktor.serialization.kotlinx.json)
     implementation(project(mapOf("path" to ":jakta-distributed:jakta-distributed-common")))
     implementation(project(mapOf("path" to ":jakta-bdi")))
-    implementation(project(mapOf("path" to ":jakta-distributed:jakta-distributed-broker")))
 }
