+++ jakta-distributed/jakta-distributed-common/src/main/kotlin/it/unibo/jakta/agents/distributed/common/Error.kt
@@ -1,0 +1,9 @@
+package it.unibo.jakta.agents.distributed.common
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class Error {
+    BAD_REQUEST,
+    CLIENT_DISCONNECTED,
+}
